+++ tests/unitary/src/compute/test_device_manager.cpp
@@ -28,33 +28,17 @@
 #include <xmipp4/core/compute/device_manager.hpp>
 
 #include <xmipp4/core/compute/device.hpp>
-#include <xmipp4/core/compute/device_backend.hpp>
 #include <xmipp4/core/compute/device_create_parameters.hpp>
 #include <xmipp4/core/version.hpp>
 
 #include <algorithm>
 
+#include "mock/mock_device_backend.hpp"
+
 #include <catch2/catch_test_macros.hpp>
-#include <trompeloeil.hpp>
 
 using namespace xmipp4;
 using namespace xmipp4::compute;
-
-class mock_device_backend final
-    : public device_backend
-{
-public:
-    MAKE_MOCK0(get_name, std::string (), const noexcept override);
-    MAKE_MOCK0(get_version, version (), const noexcept override);
-    MAKE_MOCK0(is_available, bool (), const noexcept override);
-    MAKE_MOCK0(get_priority, backend_priority (), const noexcept override);
-    MAKE_MOCK1(enumerate_devices, void (std::vector<std::size_t>&), const override);
-    MAKE_MOCK2(get_device_properties, bool (std::size_t, device_properties&), const override);
-    MAKE_MOCK2(create_device, std::unique_ptr<device> (std::size_t, const device_create_parameters&), override);
-    MAKE_MOCK2(create_device_shared, std::shared_ptr<device> (std::size_t, const device_create_parameters&), override);
-
-};
-
 
 TEST_CASE( "register device backend", "[device_manager]" ) 
 {
